Changed data types to facilitate refactoring to use camera class directly

+++ raycaster/src/camera.rs
@@ -5,9 +5,9 @@ pub struct Camera {
     pos: Vec2,
     dir: Vec2,
     plane: Vec2,
-    _fov: f32,
-    _move_speed: f32,
-    _rotation_speed: f32,
+    _fov: f64,
+    _move_speed: f64,
+    _rotation_speed: f64,
 }
 
 impl Camera {
@@ -22,38 +22,38 @@ impl Camera {
         }
     }
 
-    pub fn init(&mut self, x: f32, y: f32) {
+    pub fn init(&mut self, x: f64, y: f64) {
         self.pos.x = x;
         self.pos.y = y;
     }
 
     pub fn move_forward(&mut self, world: &World, delta_time: f64) {
-        let move_speed = self._move_speed * delta_time as f32;
+        let move_speed = self._move_speed * delta_time;
 
-        if world.is_collision((self.pos.x + self.dir.x * move_speed) as i32, self.pos.y as i32) == false {
+        if world.is_collision((self.pos.x + self.dir.x * move_speed) as i64, self.pos.y as i64) == false {
             self.pos.x += self.dir.x * move_speed;
         }
 
-        if world.is_collision(self.pos.x as i32, (self.pos.y + self.dir.y * move_speed) as i32) == false {
+        if world.is_collision(self.pos.x as i64, (self.pos.y + self.dir.y * move_speed) as i64) == false {
             self.pos.y += self.dir.y * move_speed;
         }
     }
 
     pub fn move_backward(&mut self, world: &World, delta_time: f64) {
-        let move_speed = self._move_speed * delta_time as f32;
+        let move_speed = self._move_speed * delta_time;
 
-        if world.is_collision((self.pos.x - self.dir.x * move_speed) as i32, self.pos.y as i32) == false {
+        if world.is_collision((self.pos.x - self.dir.x * move_speed) as i64, self.pos.y as i64) == false {
             self.pos.x -= self.dir.x * move_speed;
         }
 
-        if world.is_collision(self.pos.x as i32, (self.pos.y - self.dir.y * move_speed) as i32) == false {
+        if world.is_collision(self.pos.x as i64, (self.pos.y - self.dir.y * move_speed) as i64) == false {
             self.pos.y -= self.dir.y * move_speed;
         }
     }
 
     pub fn rotate_left(&mut self, delta_time: f64) {
         // Rotate the direction vector - this is effectively an imperative version of a matrix multiplication
-        let speed = self._rotation_speed * delta_time as f32;
+        let speed = self._rotation_speed * delta_time;
 
         let old_dir_x = self.dir.x;
         self.dir.x = self.dir.x * speed.cos() - self.dir.y * speed.sin();
@@ -67,7 +67,7 @@ impl Camera {
 
     pub fn rotate_right(&mut self, delta_time: f64) {
         // Rotate the direction vector - this is effectively an imperative version of a matrix multiplication
-        let speed = self._rotation_speed * delta_time as f32;
+        let speed = self._rotation_speed * delta_time;
 
         let old_dir_x = self.dir.x;
         self.dir.x = self.dir.x * -speed.cos() - self.dir.y * -speed.sin();
